Add check to scripts

+++ run_conc_model.sh
@@ -13,6 +13,13 @@ echo "target: $target"
 echo "work directory: $work"
 echo "seed directory: $seed"
 
+if [ ! -f ~/Workspace/git/drifuzz-panda/work/$target/$target.qcow2 ]; then
+    echo "Cannot run concolic script because concolic image for $target isn't setup"
+    echo "Go to drifuzz-panda directory and run:"
+    echo "  ./snapshot_helper.py $target"
+    exit 1
+fi
+
 # Prepare work directory with globalmodule.json
 rm -rf $work
 mkdir -p $work
